Refactoring STREAM event parsing
Refactoring the parsing for STREAM events (other event types are to follow).
This is related to ticket 2411 [1] where a new field will cause breakage for
our parsing. Besides that, this is both a more generic and readable parsing
implementation. Testing was by connecting to a TBB instance and listening to
STREAM events. Thanks to nickm for a rough draft of the new implementation.

[1] https://trac.torproject.org/projects/tor/ticket/2411