Refactoring ADDRMAP event parsing
Using the new style for ADDRMAP events. These events have both quoted and
non-quoted values breaking positional and keyword parsing, so keeping the regex
parsing.

This expands the event to include the ERROR and GMTExpiry attributes. This was
tested by visiting sites with TBB and printing out the ADDRMAP results.