4097 - test to see if deep copy breaks consistency
because of deepcopy here, I think that:
if refcount == 0, then TorCtl deletes the router
from its copy of the consensus, even though references
to the router are still floating around, causing keyerror.

what I think should happen:
the router should be marked as deleted but kept around until
we have no more references anywhere, and we should have
additional checks for the deleted property in PathSupport.get_exit_router()
so we can return None for those cases, and then the proper exceptions will get raised

this is confusing!